LCP/visitedPlateau removed, edgesLCP now list
visitedPlateau stored in an edge if it was already visited while
creating the plateau.
LCP was a stack which stored the visited edges to reset the
visitedPlateau value.
Both replaced by a HashSet "visited" which keeps track of the visited
edges

edgesLCP is now a list, which means less memory usage plus 2
pointer-variables less (nb_edges, Nnb_edges)
it also eliminates the parameter "M" of the methods RandomWalker, fill_A
and fill_B